requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
pymupdf==1.23.8
python-docx==0.8.11
openai==1.3.9
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
